requirements.txt
aiogram==3.20.0.post0
pydantic==2.11.7
pydantic-settings==2.10.1
PyYAML==6.0.2
python-dotenv==1.1.1
